ml/requirements.txt
librosa==0.10.2
numpy>=1.24
scipy>=1.10
soundfile>=0.12
tqdm>=4.66
PyYAML>=6.0
